captura_informacion_manual/requirements.txt
flask
python-dotenv
psycopg2-binary
pika
